# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EBOS API",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "0.0.0"
    },
    "paths": {
...
        "/api/v1/residents": {
            "get": {
                "tags": [
                    "Residents"
                ],
                "summary": "Get All Resident",
                "operationId": "getAllResident",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "format": "string",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "format": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "format": "uuid",
                                                        "example": "1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "format": "string",
                                            "example": "http://localhost:8000/api/v1/barangays?page=1"
                                        },
                                        "from": {
                                            "format": "string",
                                            "example": null
                                        },
                                        "last_page": {
                                            "format": "integer",
                                            "example": 1
                                        },
                                        "last_page_url": {
                                            "format": "string",
                                            "example": "http://localhost:8000/api/v1/barangays?page=1"
                                        },
                                        "next_page_url": {
                                            "format": "string",
                                            "example": null
                                        },
                                        "per_page": {
                                            "format": "integer",
                                            "example": 20
                                        },
                                        "total": {
                                            "format": "integer",
                                            "example": 0
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/notauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Residents"
                ],
                "summary": "Add Resident",
                "operationId": "addResident",
                "requestBody": {
                    "description": "Add Resident",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "format": "string",
                                        "example": "Juan"
                                    },
                                    "last_name": {
                                        "format": "string",
                                        "example": "Dela Cruz"
                                    },
                                    "middle_name": {
                                        "format": "string",
                                        "example": "Dela Cruz"
                                    },
                                    "address1": {
                                        "format": "string",
                                        "example": "Address 1"
                                    },
                                    "address2": {
                                        "format": "string",
                                        "example": "Address 2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns Success Message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "format": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "format": "string",
                                            "example": "Resident data"
                                        },
                                        "data": {
                                            "properties": {
                                                "resident_id": {
                                                    "format": "string",
                                                    "example": "00000-0000000-0000-00-00000"
                                                },
                                                "first_name": {
                                                    "format": "string",
                                                    "example": "Juan"
                                                },
                                                "last_name": {
                                                    "format": "string",
                                                    "example": "Dela Cruz"
                                                },
                                                "middle_name": {
                                                    "format": "string",
                                                    "example": "Dela Cruz"
                                                },
                                                "email": {
                                                    "format": "string",
                                                    "example": "example@email.com"
                                                },
                                                "contact_number": {
                                                    "format": "string",
                                                    "example": "09123456789"
                                                },
                                                "address1": {
                                                    "format": "string",
                                                    "example": "Address 1"
                                                },
                                                "address2": {
                                                    "format": "string",
                                                    "example": "Address 2"
                                                }
                                            },
                                            "type": "object",
                                            "format": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/notauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "responses": {
            "notauthenticated": {
                "description": "Respone for Unauthenticated User",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "code": {
                                    "format": "string",
                                    "example": "401"
                                },
                                "message": {
                                    "format": "string",
                                    "example": "You're unauthorized to make a request!"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}